pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "odh-runbooks"
version = "0.1.0"
description = "Verified runbook executor for ODH/RHOAI components"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
mcp = [
    "fastmcp>=2.0",
]

[project.scripts]
odh = "cli:cli"
odh-mcp = "mcp_server:mcp.run"

[tool.hatch.build.targets.wheel]
packages = ["runner", "generator", "assistant"]
